edit: utils - using size_t for buffer size

+++ src/utils.c
@@ -68,7 +68,7 @@ int getSecureInputInt(char *prompt, int bufferSize)
 }
 
 // Ask user input for a signed char : see zenithar.h
-signed char *getSecureInputSignedChar(char *prompt, int bufferSize)
+signed char *getSecureInputSignedChar(char *prompt, size_t bufferSize)
 {
   char buffer[bufferSize];
   signed char *userInput;
